fix(dflash2): context_kv_materialize 的 V 面按本 fork 约定改用 BF16（上游为 FP16）

+++ src/ops/context_kv_materialize/context_kv_materialize.cpp
@@ -95,7 +95,7 @@ void validate_cache(const CyclicKVCacheLayerView& cache, std::int32_t padded,
         throw std::invalid_argument("context_kv_materialize: invalid cyclic cache geometry");
     }
     require_tensor(cache.k, DType::BF16, kHeadDim, padded, kKVHeads, lane_capacity, 16, "cache K");
-    require_tensor(cache.v, DType::FP16, kHeadDim, padded, kKVHeads, lane_capacity, 16, "cache V");
+    require_tensor(cache.v, DType::BF16, kHeadDim, padded, kKVHeads, lane_capacity, 16, "cache V");
 }
 
 } // namespace
